build fix
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@46205 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/dbg/dbgmoddwarf.cpp
@@ -4521,7 +4521,7 @@ static DECLCALLBACK(int) rtDbgModDwarf_TryOpen(PRTDBGMODINT pMod, RTLDRARCH enmA
     pThis->aDieAllocators[1].cbMax = sizeof(RTDWARFDIECOMPILEUNIT);
     for (uint32_t i = 0; i < RT_ELEMENTS(g_aTagDescs); i++)
         if (g_aTagDescs[i].pDesc && g_aTagDescs[i].pDesc->cbDie > pThis->aDieAllocators[1].cbMax)
-            pThis->aDieAllocators[1].cbMax = g_aTagDescs[i].pDesc->cbDie;
+            pThis->aDieAllocators[1].cbMax = (uint32_t)g_aTagDescs[i].pDesc->cbDie;
     pThis->aDieAllocators[1].cbMax = RT_ALIGN_32(pThis->aDieAllocators[1].cbMax, sizeof(uint64_t));
 
     for (uint32_t i = 0; i < RT_ELEMENTS(pThis->aDieAllocators); i++)
